requirements.txt
requests
pandas
tqdm
datetime
python-dotenv
pymysql
schedule
yfinance
argparse
numpy